# atms-docs-server/api/openapi.yaml
openapi: 3.0.0
info:
  title: ATMS
  description: |-
    Attendance Tracking Management System (ATMS) is a part of an HR
    System, that allows the Admin to track the attendance and leaving of the
    employees, tracks their availability during the day, and managestheir profiles.
  contact:
    name: Ashraf Gehad
    url: https://github.com/3r3bu5
    email: deva4rf@gmail.com
  license:
    name: MIT
  version: 1.1.0
servers:
  - url: https://virtserver.swaggerhub.com/3r3bu5/ATMS/1.0.0
    description: SwaggerHub API Auto Mocking
  - url: http://localhost:3000
    description: localhost
  - url: https://localhost:9000
    description: HTTPS
security:
  - Authorization: []
...
paths:
...
  /users/{userId}/register:
    post:
      tags:
        - user
      summary: Register a new user.
      description: |-
        * Register an employee.
        * Access: This operation can only performed by Admin
      operationId: post-users-userId-register
      parameters:
        - name: userId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
        - name: Origin
          in: header
          description: CORS
          required: true
          style: simple
          explode: false
          schema:
            type: string
      requestBody:
        description: Request body will contain all the new user's information.
        content:
          application/json:
            schema:
              type: object
              properties: {}
            examples:
              example:
                value:
                  name: Ahmed El Shenawey
                  age: 25
                  password: shena1bas
                  email: shena@atms.com
                  gender: 1
                  department: 69d5b3cd-24ba-49d4-8ade-029af1530a9e
                  position: Sr. Developer
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: {}
                example: {}
                x-content-type: application/json
              examples:
                example:
                  value:
                    success: true
                    status: Registration Successful!
      security:
        - Authorization: []
      x-swagger-router-controller: User
  /users/{userId}/login:
    post:
      tags:
        - user
      summary: Login endpoint.
      description: |-
        * Login endpoint.
        * Access: This operation can only performed by Admin
      operationId: login-user
      parameters:
        - name: userId
          in: path
          required: true
          style: simple
          explode: false
          schema:
            type: string
        - name: Origin
          in: header
          description: CORS
          required: true
          style: simple
          explode: false
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties: {}
            examples:
              example:
                value:
                  email: shena@atms.com
                  password: shena1bas
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties: {}
                example: {}
                x-content-type: application/json
              examples:
                example:
                  value:
                    status: true
                    token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJfaWQiOiJmZjRlODZiOS1hNzFiLTQzNGYtYWExNi1iNWY1ZjQ1NmQzMTciLCJpYXQiOjE2MDgyOTA1MTAsImV4cCI6MTYwODMzODUxMH0.Xu5Y0KQYosW-TEO2EyqdEVZUCJ-MaqLyheowbLkqNjo
                    message: Logged-In Successful!
      security:
        - Authorization: []
      x-swagger-router-controller: User
...
components:
...
  securitySchemes:
    Authorization:
      type: http
      scheme: bearer